Cache dark mode gradient
