+++ Cargo.toml
@@ -37,3 +37,7 @@ tracing = "0.1.41"
 tracing-log = "0.2.0"
 tracing-subscriber = { version = "0.3.19", features = ["env-filter"] }
 twox-hash = { version = "2.1.1", features = ["serialize"] }
+
+[profile.bench]
+debug = true
+strip = false
